Add devices kwxx, wwvb and umc
This is an import of extensions made in my local copy of SimH 3.9 in
2016 in order to run a pre-built EPOS binary configured for the
PDP11/44 system at USC/Information Sciences Institute named ISI-MCON.
The extensions add simulations of three hardware devices installed on
that system:

- KW11-XX high-resolution clock designed and built at ISI
- Spectracom WWVB radio clock connected to a DL11 serial port
- Dummy device to stand in for an ACC UMC-Z80 interface board

The source files for the three devices were imported without change.
Additions to the remaining files were adapted for SimH 4.0.