Display GIT commit ID with simh version
Copy over some code from the full simh Makefile to store the GIT
commit ID and display it along with the simh version when simh starts
up.  Display the commit ID on a separate line since otherwise it wraps
in 80 columns with the addition of REALCONS on the version line.