Fix non-monotonic uptime
The uptime calculation based on the change in sim_gtime since reset
was producing non-monotonic results because the value needs to be
divided by sim_timer_inst_per_sec() which varies as the activity level
changes.

Now only the time between consecutive reads of the clock is calculated
based on sim_gtime and accumulated in the uint64 kwxx_uptime.