Fix flashing LEDs bug
When setting a control value from the corresponding switch register,
it was not safe to clear the value in the blinkenlight_control_t
struct first before ORing in the desired value because the separate
gpiopattern thread was reading that value for the LAMPTEST switch.
This caused the LAMPTEST switch to falsely appear set for a brief
time, so a single LED group would flash all ones.